+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Sopor Stupor Effect
 - Sopor Slime Pie food item
 - Land type extensions for datapacks, which can be used to add worldgen features, carvers and mob spawn to lands
+- New editmode settings screen, which can be used to teleport between different areas where editmode can function
 
 ### Changed
 
@@ -23,6 +24,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Fine china axe now uses decorated pot in recipe instead of flower pot
 - Changed `/setrung` command into `/rung set`
 - Minor changes and polish to the temple scanner
+- The area around which an editmode player can move has been made more flexible, each computer can act as a source to extend reach
 
 ### Fixed